Implemented native insertionsort visualization.
With bubblesort, first, I created an int[] and did one bubblesort-step on this array, then, according to the swap operation, updated (or not if no swap happened) the values and colors of the elements. This procedure screamed for optimization since this happened every single frame.
Now with insertionsort, the sorting algorithm step operates natively on an Element[]. No more for-looping to get all values of the elements in an int[] or updating those values.