Implemented selectionsort visualization.
Element which currently is thought to be the minimum is marked and
element which currently is compared with minimum is marked. If
compareIndex reaches end of array, the minimum found will be swapped
with first unsorted element. Then the minimum will be marked as sorted
with a green transparent background.