Reason for faulty visualization found.
Mergesort does his job well but the swaps only consider the values of
the elements, not the position. If two elements swap, they should only
swap their value and color, not everything since this results in the same element
in the array, only on a different index. This does no difference for the
visualization.

In a nutshell, a Element.swap(Element) function is needed.