Implemented mergesort for a set of elements.
Implementing this was not that easy since merging two sets together
sometimes didn't return a sorted list since values AND coordinates have
to be sorted for a set of elements which wasn't always the case.
The element with the lowest value must also have the lowest x-coordinate
to be the most left element. After comparing two elements, they would swap
if the left element would have a higher value than the right. Swapping means,
they would swap their x-coordinates. After this, the right element with
the new x-coordinate would be added to the merged list. But now the
order of the both remaining elements in the subsets is no longer
guaranteed. Both subsets have now to be resorted before further merging.
This is done in the function restoreoOrder(Element[], int). Since only
after the swapped element the subset is no longer sorted, passing the
index to the function improves performance since not the whole subset's
order has to be checked only after the passed index.