Implemented very simple sonification of bubblesort.
When a swap happened, the synth defined in the file 'bubblesort.sc' will
play with a pitch between 200 and 1200. The absolute difference between
the elements involved in the swap will define the pitch. This version
should only be played with a framerate lower than 15. Otherwise, the
synths will create too much interference.