Overhauled bubblesort visualization.
Previous visualization was very unintuitive. Updated visualization now
uses a thread in which bubblesort gets executed and after every logical frame
the canvas gets notified to draw the new frame. (Almost) All
of the codebase has been changed during this overhaul. Implementation of
the other sorting algorithms will be adjusted in the future to match this
design of implementation.