Mark for cut index is now between elements.
Changed visual for the element at cut index. The red marker is now
between the cut index and its left element.
Implemented by setting the state of the left element as marked (not the
element at cut index) and drawing the marker to its right. Since this
way the marker gets drawn first, the marker will not hide any elements in the
background.