Changed parameters from Mergesort-constructor.
In the main-file, you no longer have to create all the objects like
the elements and the int-array. Creating a instance of the class
Mergesort will handle all of this by itself. Only parameter now is
amount of elements to sort.
Furthermore, the lock object was redundant since you can just use the
Mergesort-instance for synchronizing. That also does make a lot more
sense. To be able to implement all of this this way, the method to
create elements has been renamed from 'getElements' to 'createElements'.
Parameters have also changed here. No color[] needed anymore. Most of
the times (all), I would use the standard color array anyway, so passing it
was not a clever way of handling things.