Added multichannel expansion and more to synth.
The given frequency now is expanded into 4 channels whereas the lowest
frequency has an amplitude of 1 and the highest frequency an amplitude
of 0.7 times a modifier which depends on the base frequency. The higher
the frequency, the lower the modifier. The resulting channels are mixed
into one stereo-channel.

Changed frequency spectrum from 200-1640 to 200-4000.