Paths to OSCdef and names of synths are now unique.
This prevents overriding the listener and synths of different sorting
algorithms when using the same sonification variant. Every OSCdef and
every synth is now unique and will no longer be overriden when executing
SuperCollider code of a different file.

Also unified UI from Quicksort and Selectionsort with the other
algorithms. (Change-button was missing)