Boot sound also plays after first received status-message.
Previously, if the user did not evaluate the .sc-file before starting
the sketch, the sketch would indicate this with the red status-icon. When
evaluating, the icon would become green but no boot sound would play.
With this commit, a boot sound will play when connection has been
established.

If the file is already evaluated before the start of the sketch, two
boot sounds will play simultaneously. One because of the boot-message
and a second one because of the first status-message. This seems to not
be a great issue but can and probably will be fixed in future commits.