Implemented specific sonifications.
Sonification are now algorithm-specific. This means that the
sonifications of each algorithm is now implemented inside the algorithm
and not outside in a Sonification-enum. Switching between sonifications
is now implemented in the abstract class SortingThread. To add a
sonification to a algorithm, you have to add the Sonification-instance
to the list of sonifications of the super-class. You can then switch
between those sonifications via the "SONI"-button. Resetting an
algorithm resets the elements but remembers the selected sonification.
Adding the remaining algorithms including their sonifications should now
be easy.