Fixed bug: Changing algorithm -/-> Algorithm is reset
Encountered a bug, where pausing a algorithm, resetting it, and then
changing back to the same algorithm leads to a pause screen. The issue
was that resets were handled independent of the algorithm list in
AlgosoundUtil. When changing algorithms, this list got accessed and all
the algorithms inside it are never reset.

Fixed with new method updateAlgorithm() in AlgosoundUtil. It overwrites
the instances in the list with the new given instances (which are
reset).